Extract convertToVisibleTags into AmenityTagsService
